+++ hash.go
@@ -95,57 +95,28 @@ const (
 )
 
 
-func bufferSize(fileSize int64) int {
-    switch {
-    case fileSize < 10*MB:
-        return 32 * KB
-    case fileSize < 1024*MB:
-        return 1 * MB
-    default:
-        return 5 * MB
-    }
-}
-
 // hash a file with moist effective buffer size
+// TODO : memory allocation improvement
 func HashFile(hash, filePath string) (string, error) {
-    fi, err := os.Stat(filePath)
-    if err != nil {
-        return "", fmt.Errorf("stat file: %w", err)
-    }
-
-    bufSize := bufferSize(fi.Size())
-
     file, err := os.Open(filePath)
     if err != nil {
         return "", fmt.Errorf("open file: %w", err)
     }
     defer file.Close()
 
-    factory, ok := Registry[strings.ToLower(hash)]
+    constructor, ok := Registry[strings.ToLower(hash)]
     if !ok {
         return "", fmt.Errorf("unsupported hash: %s", hash)
     }
 
-    hasher := factory()
-    buf := make([]byte, bufSize)
-
-    for {
-        n, err := file.Read(buf)
-        if n > 0 {
-            if _, werr := hasher.Write(buf[:n]); werr != nil {
-                return "", fmt.Errorf("hash write: %w", werr)
-            }
-        }
-        if err == io.EOF {
-            break
-        }
-        if err != nil {
-            return "", fmt.Errorf("read file: %w", err)
-        }
-    }
+	hasher := constructor()
+	if _, err := io.Copy(hasher, file); err != nil {
+		return "", err
+	}
 
     return hex.EncodeToString(hasher.Sum(nil)), nil
 }
+
 
 func HashFileKey(hash string, key []byte, filePath string) (string, error) {
     file, err := os.Open(filePath)